feat: generate-ai-pick-animation
